extract writing the Vagrantfile into a helper module and add an :around hook for the tempdir creation
